fix(migration): move SearchOverlay to packages/ui, restore deleted components, fix transparent overlay
- Move SearchOverlay from apps/expo/components/ to packages/ui/src/search-overlay/
- Fix transparent iOS overlay by adding bg-background to the absoluteFill Animated.View
- Restore LargeTitleHeaderSearchContentContainer in packages/ui/src/ (provides opaque bg + SafeAreaView)
- Restore LargeTitleHeaderOverlapFixIOS in packages/ui/src/ (provides iOS header overlap fix)
- Fix dashboard empty-state text: revert t('dashboard.searchPrompt') back to t('dashboard.searchPlaceholder') = "Search dashboard"
- Remove incorrectly added dashboard.searchPrompt translation key from en.json
- Update all 6 screen imports to use @packrat/ui/src/search-overlay
- Update packages/ui/tsconfig.json to include src/**/*
- Correct migration doc: components should be moved, not deleted
