+++ apps/expo/lib/i18n/locales/en.json
@@ -217,8 +217,7 @@
     "title": "Dashboard",
     "searchPlaceholder": "Search dashboard",
     "noResults": "No matching tiles found",
-    "tryDifferent": "Try different keywords or clear your search",
-    "searchPrompt": "Type to find tiles on your dashboard"
+    "tryDifferent": "Try different keywords or clear your search"
   },
   "packs": {
     "currentPack": "Current Pack",
