+++ packages/ui/src/large-title-header-overlap-fix-ios.tsx
@@ -1,0 +1,15 @@
+import type { ReactNode } from 'react';
+import { Platform, SafeAreaView, View } from 'react-native';
+
+export function LargeTitleHeaderOverlapFixIOS({ children }: { children?: ReactNode }) {
+  if (Platform.OS === 'android') {
+    if (!children) return null;
+    return <>{children}</>;
+  }
+  return (
+    <SafeAreaView {...(children ? { className: 'flex-1 bg-background' } : {})}>
+      {children}
+      {!children && <View className="h-[0.5]" />}
+    </SafeAreaView>
+  );
+}
